third_party/WebKit/LayoutTests/web-animations-api/player-expected.txt

This is a testharness.js-based test.
FAIL Player sanity tests Cannot read property 'player' of undefined
PASS Player.finish() raises exception if source content end is infinity 
PASS Player.startTime/currentTime/playbackRate throw if passed non-finite numbers. 
Harness: the test ran to completion.
